Rework of formGenerator to use a component Allow HTML for label & hint
